Move attributes, read/write
